Little changes to make testing simpler
Moved ec2metadata into a standalone function and the ec2metadata
service does not need to be instantiated as part of Cloud, makes it
simpler to test.
