+++ pkg/cloudprovider/providers/aws/node.go
@@ -1,13 +1,12 @@
 package aws
 
 import (
-	"io/ioutil"
-
 	"github.com/UKHomeOffice/keto/pkg/cloudprovider"
 	"github.com/UKHomeOffice/keto/pkg/model"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go/aws/ec2metadata"
+	"github.com/aws/aws-sdk-go/aws/session"
 )
 
 // Node returns an implementation of Node interface for AWS Cloud.
@@ -17,36 +16,41 @@ func (c *Cloud) Node() (cloudprovider.Node, bool) {
 
 // GetKubeAPIURL returns a full URL to Kubernetes API.
 func (c *Cloud) GetKubeAPIURL() (string, error) {
-	instanceID, err := c.getInstanceID()
+	metadata, err := getInstanceMetadata()
 	if err != nil {
 		return "", err
 	}
-	return c.getResourceTagValue(instanceID, kubeAPIURLTagKey)
+	return c.getResourceTagValue(metadata.InstanceID, kubeAPIURLTagKey)
 }
 
 // GetKubeVersion returns a kubernetes version string.
 func (c *Cloud) GetKubeVersion() (string, error) {
-	instanceID, err := c.getInstanceID()
+	metadata, err := getInstanceMetadata()
 	if err != nil {
 		return "", err
 	}
-	return c.getResourceTagValue(instanceID, kubeVersionTagKey)
+	return c.getResourceTagValue(metadata.InstanceID, kubeVersionTagKey)
 }
 
-// getInstanceID returns an instance ID from EC2 metadata service.
-func (c Cloud) getInstanceID() (string, error) {
-	return c.ec2Metadata.GetMetadata("instance-id")
+// getInstanceMetadata returns an instance metadata document from EC2 metadata service.
+func getInstanceMetadata() (ec2metadata.EC2InstanceIdentityDocument, error) {
+	sess, err := session.NewSession(&aws.Config{})
+	if err != nil {
+		return ec2metadata.EC2InstanceIdentityDocument{}, err
+	}
+	metadataService := ec2metadata.New(sess)
+	return metadataService.GetInstanceIdentityDocument()
 }
 
 // GetAssets gets assets onto a filesystem.
 func (c *Cloud) GetAssets() (model.Assets, error) {
 	a := model.Assets{}
 
-	instanceID, err := c.getInstanceID()
+	metadata, err := getInstanceMetadata()
 	if err != nil {
 		return a, err
 	}
-	bucket, err := c.getResourceTagValue(instanceID, assetsBucketNameTagKey)
+	bucket, err := c.getResourceTagValue(metadata.InstanceID, assetsBucketNameTagKey)
 	if err != nil {
 		return a, err
 	}
@@ -74,23 +78,4 @@ func (c *Cloud) GetAssets() (model.Assets, error) {
 	a.KubeCACert = kubeCACert
 
 	return a, nil
-}
-
-func (c Cloud) getS3Object(bucket, objectName string) ([]byte, error) {
-	params := &s3.GetObjectInput{
-		Bucket: aws.String(bucket),
-		Key:    aws.String(objectName),
-	}
-
-	resp, err := c.s3.GetObject(params)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	return b, nil
 }
